requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic[email]
pydantic-settings
python-jose[cryptography]
bcrypt
python-multipart
httpx
pytest
pytest-asyncio
aiosqlite
